Fix GitHub #7: Problem in OrderedMap.delete
Several parts of the code relied on the falsy-ness of `nil` when
evaluating the value of `map[key]`, as in `if map[key], do: ...`. This
led to erroneous behavior when map values were boolean. Specifically, a
value of `false` would cause `map[key]` to be interpreted as indicating
that the given key did not exist, even when it clearly did.

All instances of the above usage have been replaced with
`Map.has_key?/2`. Tests for all relevant pathological cases have also
been added.